First batch adapting to the new spec
