+++ .github/.jira_sync_config.yaml
@@ -5,10 +5,7 @@
 settings:
   # Repository specific settings
   components: # Jira components that will be added to Jira issue
-    - mysql-vm
-    - mysql-k8s
-    - postgresql-vm
-    - postgresql-k8s
+    - opensearch-vm
 
   # Settings shared across Data Platform repositories
   label_mapping:
