+++ metadata.yaml
@@ -3,15 +3,15 @@
 name: opensearch-benchmark
 display-name: Charmed OpenSearch Benchmark
 
-summary: Generates a load against your workload and collect performance metrics to Prometheus.
+summary: Generates load against your workload and collect performance metrics to Prometheus.
 description: |
   Run this charm to connect with an opensearch deployment and test the performance or generate a load against the cluster.
 
 # TODO: Update
-docs: https://discourse.charmhub.io/t/charmed-sysbench-documentation-home/13945
-source: https://github.com/canonical/sysbench-operator
-issues: https://github.com/canonical/sysbench-operator/issues
-website: https://charmhub.io/sysbench
+# docs: https://discourse.charmhub.io/t/charmed-sysbench-documentation-home/13945
+source: https://github.com/canonical/opensearch-benchmark-operator
+issues: https://github.com/canonical/opensearch-benchmark-operator/issues
+# website: https://charmhub.io/opensearch-benchmark
 maintainers:
   - Canonical Data Platform <data-platform@lists.launchpad.net>
