+++ pyproject.toml
@@ -5,34 +5,34 @@
 # Charm is not packed as a standard Python package; this information is not used
 name = "charm"
 version = "0.1.0"
-description = ""
-authors = []
+description = "Generates load against your workload and collect performance metrics to Prometheus."
+authors = ["Canonical Data Platform <data-platform@lists.launchpad.net>"]
 
 [tool.poetry.dependencies]
 python = "^3.10"
 ops = "^2.17.0"
-tenacity = "^8.5.0"
+tenacity = "^9.0.0"
 jinja2 = "^3.1.4"
 overrides = "7.7.0"
 requests = "2.32.3"
 shortuuid = "1.0.13"
-pydantic = "^1.10.17, <2"
+pydantic = "^1.10.18, <2"
 cryptography = "^43.0.1"
 jsonschema = "^4.23.0"
 
 
 [tool.poetry.group.charm-libs.dependencies]
 # data_platform_libs/v0/data_interfaces.py
-ops = ">=2.0.0"
+ops = "^2.17"
 # data_platform_libs/v0/upgrade.py
 # grafana_agent/v0/cos_agent.py requires pydantic <2
 pydantic = "^1.10, <2"
 # tls_certificates_interface/v1/tls_certificates.py
 cryptography = "^43.0.0"
-jsonschema = "^4.21.1"
+jsonschema = "^4.23.0"
 # grafana_agent/v0/cos_agent.py
-cosl = ">=0.0.7"
-bcrypt = ">=4.0.1"
+cosl = "^0.0.41"
+bcrypt = "^4.1.3"
 
 [tool.poetry.group.format]
 optional = true
@@ -49,7 +49,7 @@ shellcheck-py = "^0.10.0.1"
 
 [tool.poetry.group.unit.dependencies]
 pytest = "^8.3.3"
-pytest-asyncio = "*"
+pytest-asyncio = "^0.21.2"
 coverage = {extras = ["toml"], version = "^7.6.1"}
 parameterized = "^0.9.0"
 
@@ -62,7 +62,7 @@ pytest-operator-groups = {git = "https://github.com/canonical/data-platform-work
 pytest-microceph = {git = "https://github.com/canonical/data-platform-workflows", tag = "v22.0.0", subdirectory = "python/pytest_plugins/microceph"}
 juju = "^3.5.2"
 ops = "^2.17.0"
-tenacity = "^8.5.0"
+tenacity = "^9.0.0"
 pyyaml = "^6.0.2"
 urllib3 = "^2.2.3"
 protobuf = "5.28.2"
